pyproject.toml
[tool.poetry]
name = "civ-weave"
version = "0.1.0"
description = ""
authors = ["James McNeill <jpmmcneill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.5.1"
kuzu = "^0.1.0"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
pytest = "^7.4.3"
mypy = "^1.7.1"
yamllint = "^1.33.0"


[tool.ruff]
select = ["ALL"]
line-length = 79
indent-width = 4
ignore = ["D203", "D212", "ISC001"]
# D203 and D212 lead to ugly docstrings IMO
# ISC001 ignored to get rid of a conflict warning (even though it's a good rule)

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
